FEM: fem tools package, move ccx tools in it
